requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
psycopg[binary]==3.1.19
